chore: remove Icon component
