+++ libs/design-system-core/src/components/ExtensionIcon/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { AkashaAppApplicationType } from '@akashaorg/typings/lib/sdk/graphql-types-new';
 import { Layers2Icon } from 'lucide-react';
-import Icon, { IconProps } from '../Icon';
+import { IconProps } from '../AppIcon';
 import { Plugin, Widget } from '../Icon/akasha-icons';
 
 export type ExtensionIconProps = IconProps & {
@@ -15,7 +15,9 @@ const ExtensionIcon: React.FC<ExtensionIconProps> = props => {
     solid = false,
     accentColor = true,
     type,
-    defaultIcon = <Layers2Icon className="h-5 w-5 [&>*]:stroke-secondaryLight dark:[&>*]:stroke-secondaryDark" />,
+    defaultIcon = (
+      <Layers2Icon className="h-5 w-5 [&>*]:stroke-secondaryLight dark:[&>*]:stroke-secondaryDark" />
+    ),
   } = props;
 
   const getIconByType = (_type: AkashaAppApplicationType) => {
@@ -31,16 +33,7 @@ const ExtensionIcon: React.FC<ExtensionIconProps> = props => {
     }
   };
 
-  return (
-    <Icon
-      size={size}
-      solid={
-        [AkashaAppApplicationType.Plugin, AkashaAppApplicationType.Widget].includes(type) || solid
-      }
-      accentColor={accentColor}
-      icon={getIconByType(type)}
-    />
-  );
+  return getIconByType(type);
 };
 
 export default ExtensionIcon;
