--- libs/design-system-core/src/components/Icon/hero-icons-solid.tsx
@@ -1,1 +1,0 @@
-export * from '@heroicons/react/24/solid';
