--- libs/design-system-core/src/components/Icon/index.tsx
@@ -1,113 +1,0 @@
-import React from 'react';
-
-import Stack from '../Stack';
-
-import { BasicSize, BreakPointSize } from '../types/common.types';
-import { cn } from '@akashaorg/ui/lib/library/utils';
-
-export interface IconProps {
-  ref?: React.Ref<HTMLDivElement>;
-  icon?: React.ReactElement;
-  size?: BasicSize;
-  breakPointSize?: BreakPointSize;
-  accentColor?: boolean;
-  disabled?: boolean;
-  dataTestId?: string;
-  customStyle?: string;
-  solid?: boolean;
-  rotateAnimation?: boolean;
-}
-
-/**
- * An Icon's purpose is as clear as its name indicates: provide an icon. In our design
- * system, we use [Heroicons](https://heroicons.com/) together with our own icons to make
- * up a rich icon library that you can easily include in your app and apply custom styling
- * to them through props as you see fit.
- * @param color - (optional) customize the color of your icon
- * @param ref - (optional) pass the ref here
- * @param size - (optional) customize the size of your icon
- * @param breakPointSize - (optional) specify breakpoint sizes if needed
- * @param accentColor - boolean (optional) whether your icon will have the default accent color scheme
- * @param disabled - boolean (optional) a disabled icon looks different
- * @param dataTestId - (optional) useful when writing tests for the component
- * @param customStyle - (optional) apply your custom styling (Make sure to use standard Tailwind classes)
- * @param solid - boolean (optional) a solid icon looks different from an outlined one
- * @param rotateAnimation - boolean (optional) for those who want to add rotate animation to their icon
- * @example
- * ```tsx
- *  <Icon accentColor icon={<XMarkIcon />} />
- * ```
- **/
-const Icon: React.FC<IconProps> = props => {
-  const {
-    icon,
-    ref,
-    accentColor = false,
-    size = 'md',
-    breakPointSize,
-    disabled,
-    dataTestId,
-    customStyle = '',
-    solid = false,
-    rotateAnimation = false,
-  } = props;
-
-  const breakPointStyle = breakPointSize
-    ? ICON_SIZE_MAP_BY_BREAKPOINT(breakPointSize.breakPoint)[breakPointSize.size]
-    : '';
-  const sizeStyle = `${ICON_SIZE_MAP[size]} ${breakPointStyle}`;
-
-  const baseStyle = `select-none`;
-
-  const accentColorStyle = accentColor
-    ? `${
-        solid
-          ? '[&>*]:fill-secondaryLight dark:[&>*]:fill-secondaryDark'
-          : '[&>*]:stroke-secondaryLight dark:[&>*]:stroke-secondaryDark'
-      }`
-    : '';
-
-  let colorStyle: string;
-  if (!customStyle.includes('stroke') && !customStyle.includes('fill')) {
-    colorStyle = solid
-      ? '[&>*]:fill-black dark:[&>*]:fill-white'
-      : '[&>*]:stroke-secondaryLight dark:[&>*]:stroke-secondaryDark';
-  }
-
-  const activeIconColorStyle = accentColor ? accentColorStyle : colorStyle;
-
-  const disabledStyle = disabled ? 'opacity-50' : '';
-
-  const rotateStyle = rotateAnimation ? 'animate-spin' : '';
-
-  const iconStyle = cn(
-    `${baseStyle} ${activeIconColorStyle} ${sizeStyle} ${disabledStyle} ${rotateStyle}`,
-  );
-
-  return (
-    <Stack ref={ref} customStyle={customStyle}>
-      {React.cloneElement(icon, {
-        className: iconStyle,
-        'data-testid': dataTestId,
-      })}
-    </Stack>
-  );
-};
-
-const ICON_SIZE_MAP: Record<BasicSize, string> = {
-  xs: 'h-3 w-3',
-  sm: 'h-4 w-4',
-  md: 'h-5 w-5',
-  lg: 'h-6 w-6',
-  xl: 'h-7 w-7',
-};
-
-const ICON_SIZE_MAP_BY_BREAKPOINT = (breakPoint: string): Record<BasicSize, string> => ({
-  xs: `${breakPoint}:h-3 ${breakPoint}:w-3`,
-  sm: `${breakPoint}:h-4 ${breakPoint}:w-4`,
-  md: `${breakPoint}:h-5 ${breakPoint}:w-5`,
-  lg: `${breakPoint}:h-6 ${breakPoint}:w-6`,
-  xl: `${breakPoint}:h-7 ${breakPoint}:w-7`,
-});
-
-export default Icon;
